requirements.txt
staroid
findspark
python3-wget